stat and longer ass to master

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,9 +25,9 @@ public class SalesManager {
         return min;
     }
 
-    public int medium() {
-        int medium = 0;
-        for (int sale : sales) {
+    public long medium() {
+        long medium = 0;
+        for (long sale : sales) {
             medium = medium + sale;
         }
         medium = (medium - (this.max() + this.min())) / (this.sales.length - 2);
